+++ src/_rules/flex.js
@@ -1,7 +1,6 @@
-import { handler as h } from '#utils';
+import { handler as h, resolveArbitraryValues } from '#utils';
 
 export const flex = [
-  // flex
   [
     /^flex-(.*)$/,
     ([_, d]) => ({ flex: h.bracket(d) != null ? h.bracket(d).split(' ').map(e => h.cssvar.fraction(e) ?? e).join(' ') : h.cssvar.fraction(d) }),
@@ -10,15 +9,17 @@ export const flex = [
   ['flex-auto', { flex: '1 1 auto' }],
   ['flex-initial', { flex: '0 1 auto' }],
   ['flex-none', { flex: 'none' }],
-  // shrink/grow/basis
-  // grow and shrink have two forms - 'shrink' and 'shrink-N'
+
+  // shrink / grow / basis
+  ['shrink', { 'flex-shrink': 1 }],
   [
-    /^shrink(?:-(.*))?$/,
-    ([, d = '']) => ({ 'flex-shrink': h.number(d) ?? 1 }),
+    /^shrink-\[?(\d+)]?$/,
+    ([, d]) => ({ 'flex-shrink': h.number(d) ?? 1 }),
     { autocomplete: ['shrink-<num>'] },
   ],
+  ['grow', { 'flex-grow': 1 }],
   [
-    /^grow(?:-(.*))?$/,
+    /^grow-\[?(\d+)]?$/,
     ([, d = '']) => ({ 'flex-grow': h.number(d) ?? 1 }),
     { autocomplete: ['grow-<num>'] },
   ],
@@ -27,12 +28,21 @@ export const flex = [
     ([, d], { theme }) => ({ 'flex-basis': theme.spacing?.[d] ?? h.auto.fraction(d) }),
     { autocomplete: ['basis-$spacing'] },
   ],
-  // directions
+  // matching arbitrary values
+  [
+    /^basis-\[(.\d*)(rem|px|%)?]$/,
+    ([, value, unit], context) => ({
+      'flex-basis': resolveArbitraryValues(value, unit, context),
+    }),
+  ],
+
+  // direction
   ['flex-row', { 'flex-direction': 'row' }],
   ['flex-row-reverse', { 'flex-direction': 'row-reverse' }],
   ['flex-col', { 'flex-direction': 'column' }],
   ['flex-col-reverse', { 'flex-direction': 'column-reverse' }],
-  // wraps
+
+  // wrap
   ['flex-wrap', { 'flex-wrap': 'wrap' }],
   ['flex-wrap-reverse', { 'flex-wrap': 'wrap-reverse' }],
   ['flex-nowrap', { 'flex-wrap': 'nowrap' }],