+++ test/focus-ring.js
@@ -16,10 +16,10 @@ test("focus ring", async (t) => {
 
   expect(css).toMatchInlineSnapshot(`
     "/* layer: default */
-    .focus-within\\\\:focusable:focus-within{outline:2px solid var(--w-s-color-border-focused, var(--w-s-color-focused));outline-offset:var(--w-outline-offset, 1px);}
+    .focus-within\\:focusable:focus-within{outline:2px solid var(--w-s-color-border-focused, var(--w-s-color-focused));outline-offset:var(--w-outline-offset, 1px);}
     .focusable:focus,.focusable:focus-visible{outline:2px solid var(--w-s-color-border-focused, var(--w-s-color-focused));outline-offset:var(--w-outline-offset, 1px);}.focusable:not(:focus-visible){outline:none;}
-    .group:focus .group-focus\\\\:focusable,.group:focus-visible .group-focus\\\\:focusable{outline:2px solid var(--w-s-color-border-focused, var(--w-s-color-focused));outline-offset:var(--w-outline-offset, 1px);}.group:not(:focus-visible) .group-focus\\\\:focusable{outline:none;}
-    .peer:focus~.peer-focus\\\\:focusable,.peer:focus-visible~.peer-focus\\\\:focusable{outline:2px solid var(--w-s-color-border-focused, var(--w-s-color-focused));outline-offset:var(--w-outline-offset, 1px);}.peer:not(:focus-visible)~.peer-focus\\\\:focusable{outline:none;}
+    .group:focus .group-focus\\:focusable,.group:focus-visible .group-focus\\:focusable{outline:2px solid var(--w-s-color-border-focused, var(--w-s-color-focused));outline-offset:var(--w-outline-offset, 1px);}.group:not(:focus-visible) .group-focus\\:focusable{outline:none;}
+    .peer:focus~.peer-focus\\:focusable,.peer:focus-visible~.peer-focus\\:focusable{outline:2px solid var(--w-s-color-border-focused, var(--w-s-color-focused));outline-offset:var(--w-outline-offset, 1px);}.peer:not(:focus-visible)~.peer-focus\\:focusable{outline:none;}
     .focusable-inset{--w-outline-offset:-3px;}"
   `);
 });